Separate url input component
